Handle legacy doc format
